scripts/pyproject.toml
[project]
name = "llmkg-streamlit"
version = "0.1.0"
description = "Streamlit POC clients for the LLMKG project"
authors = [
    {name = "hsadler", email = "harrysadlermusic@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.99.1",
    "streamlit>=1.33.0",
    "graphviz>=0.20.3",
    "python-dotenv>=1.0.0",
    "fire>=0.7.1",
]
